Introducing GitHub Action workflows for automatic versioning, release, and tests
A set of three workflows were introduced to streamline the versioning, release, and testing process of this project. The first workflow involves automatically bumping the version with every push made to the main branch. The second workflow provides functionality for creating and publishing a release upon the creation of a new tag. Lastly, the third workflow ensures Python tests are performed on all pushes or pull requests made to the main branch. A line was also added to setup.cfg to track version numbers.