Update github actions user for automated version tagging
The user for the automated version bumping in the github actions was changed from the generic "action@github.com" to a specific user "khiladisngh@github.com". This change allows for more accountability in case of debugging and also reflects the actual user performing the action instead of a generic GitHub action user. The 'build' job was also renamed to 'version' for better understanding of its purpose.