Add GitHub Actions for automatic releases
A new GitHub Actions workflow 'release.yml' was added for auto-generating release notes and creating GitHub releases whenever a new tag is pushed. The 'tag.yml' workflow was also modified to remove the 'test' job as it's now not necessary in the context of this change. 'generate_release_notes.py' script now handles potential exceptions and checks the directory is a Git repository before proceeding. The directory structure was also updated to move this script into a 'scripts' directory. Lastly, the file specified in '.bumpversion.cfg' was corrected from '.bumpversion.cfg' to 'setup.cfg' to correctly reflect the configuration of version bumps.